requirements.txt
requests==2.25.1
pandas==1.2.3
psycopg2-binary==2.8.6
SQLAlchemy==1.3.23
redshift_connector==2.0.882